+++ Source/CesiumRuntime/Private/Tests/CesiumFeatureIdAttribute.spec.cpp
@@ -126,7 +126,7 @@ void FCesiumFeatureIdAttributeSpec::Define() {
     });
   });
 
-  Describe("GetVertexCount", [this]() {
+  Describe("GetCount", [this]() {
     BeforeEach([this]() {
       model = CesiumGltf::Model();
       CesiumGltf::Mesh& mesh = model.meshes.emplace_back();
@@ -149,7 +149,7 @@ void FCesiumFeatureIdAttributeSpec::Define() {
           ECesiumFeatureIdAttributeStatus::ErrorInvalidAccessor);
       TestEqual(
           "VertexCount",
-          UCesiumFeatureIdAttributeBlueprintLibrary::GetVertexCount(
+          UCesiumFeatureIdAttributeBlueprintLibrary::GetCount(
               featureIDAttribute),
           0);
     });
@@ -177,13 +177,13 @@ void FCesiumFeatureIdAttributeSpec::Define() {
           ECesiumFeatureIdAttributeStatus::Valid);
       TestEqual(
           "VertexCount",
-          UCesiumFeatureIdAttributeBlueprintLibrary::GetVertexCount(
+          UCesiumFeatureIdAttributeBlueprintLibrary::GetCount(
               featureIDAttribute),
           vertexCount);
     });
   });
 
-  Describe("GetFeatureIDForVertex", [this]() {
+  Describe("GetFeatureID", [this]() {
     BeforeEach([this]() {
       model = CesiumGltf::Model();
       CesiumGltf::Mesh& mesh = model.meshes.emplace_back();
@@ -206,7 +206,7 @@ void FCesiumFeatureIdAttributeSpec::Define() {
           ECesiumFeatureIdAttributeStatus::ErrorInvalidAccessor);
       TestEqual(
           "FeatureIDForVertex",
-          UCesiumFeatureIdAttributeBlueprintLibrary::GetFeatureIDForVertex(
+          UCesiumFeatureIdAttributeBlueprintLibrary::GetFeatureID(
               featureIDAttribute,
               0),
           -1);
@@ -234,13 +234,13 @@ void FCesiumFeatureIdAttributeSpec::Define() {
           ECesiumFeatureIdAttributeStatus::Valid);
       TestEqual(
           "FeatureIDForNegativeVertex",
-          UCesiumFeatureIdAttributeBlueprintLibrary::GetFeatureIDForVertex(
+          UCesiumFeatureIdAttributeBlueprintLibrary::GetFeatureID(
               featureIDAttribute,
               -1),
           -1);
       TestEqual(
           "FeatureIDForOutOfBoundsVertex",
-          UCesiumFeatureIdAttributeBlueprintLibrary::GetFeatureIDForVertex(
+          UCesiumFeatureIdAttributeBlueprintLibrary::GetFeatureID(
               featureIDAttribute,
               10),
           -1);
@@ -269,7 +269,7 @@ void FCesiumFeatureIdAttributeSpec::Define() {
       for (size_t i = 0; i < featureIDs.size(); i++) {
         TestEqual(
             "FeatureIDForVertex",
-            UCesiumFeatureIdAttributeBlueprintLibrary::GetFeatureIDForVertex(
+            UCesiumFeatureIdAttributeBlueprintLibrary::GetFeatureID(
                 featureIDAttribute,
                 static_cast<int64>(i)),
             featureIDs[i]);